Module Context.Client_state

type t

val find : Tezos_client_008_PtEdo2Zk.Protocol_client_context.full -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> t -> (Contract_state.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val register : Tezos_client_008_PtEdo2Zk.Protocol_client_context.full -> force:bool -> default_memo_size:int option -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Tezos_sapling.Core.Client.Viewing_key.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val sync_and_scan : Tezos_client_008_PtEdo2Zk.Protocol_client_context.full -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Contract_state.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Synchronise our local state with the blockchain's. The state must be recent enough to craft correct transactions. The limit enforced by the protocol if 120 blocks. Also scans, ie. checks for incoming payments and add them to our balance. *